add windows bridge format
when using podman-remote on windows, the bridge format must account for
how windows deals with escape quoting.  in this case, it does not need
any.

also,  reduced duplicated code around generating the bridge endpoint for
the unix and windows platforms.
