+++ pkg/adapter/client_windows.go
@@ -3,13 +3,13 @@
 package adapter
 
 import (
+	"fmt"
+
 	"github.com/containers/libpod/cmd/podman/remoteclientconfig"
-	"github.com/containers/libpod/libpod"
 )
 
-func newBridgeConnection(formattedBridge string, remoteConn *remoteclientconfig.RemoteConnection, logLevel string) (*Endpoint, error) {
-	// TODO
-	// Unix and Windows appear to quote their ssh implementations differently therefore once we figure out what
-	// windows ssh is doing here, we can then get the format correct.
-	return nil, libpod.ErrNotImplemented
+func formatDefaultBridge(remoteConn *remoteclientconfig.RemoteConnection, logLevel string) string {
+	return fmt.Sprintf(
+		`ssh -T %s@%s -- /usr/bin/varlink -A '/usr/bin/podman --log-level=%s varlink $VARLINK_ADDRESS' bridge`,
+		remoteConn.Username, remoteConn.Destination, logLevel)
 }
